refactor(app): extract process_async_messages dispatch (phase 6b)
Move process_async_messages out of mod.rs into a new
app/async_dispatch.rs sibling. ~660 lines.

This is the per-frame async-message orchestrator: it drains the
AsyncBridge queue and routes each AsyncMessage variant to the right
handler — LSP responses, plugin commands, filesystem polling, git
status, and so on. The match has dozens of arms, most one-line
delegations to specialized handlers in sibling files.

mod.rs goes from 7218 -> 6561 lines (-657). 2026 tests still pass.
